add: LocationGateway
