+++ client_dashboard/common/src/commonMain/kotlin/org/thechance/common/data/remote/gateway/location_gateway/ILocationGateway.kt
@@ -1,0 +1,7 @@
+package org.thechance.common.data.remote.gateway.location_gateway
+
+import org.thechance.common.domain.entity.Location
+
+interface ILocationGateway {
+    suspend fun getCurrentLocation(): Location
+}
