+++ client_dashboard/common/src/commonMain/kotlin/org/thechance/common/data/remote/gateway/user_gateway/UsersGateway.kt
@@ -1,15 +1,24 @@
 package org.thechance.common.data.remote.gateway.user_gateway
 
 import io.ktor.client.HttpClient
+import io.ktor.client.call.body
+import io.ktor.client.plugins.ClientRequestException
 import io.ktor.client.request.forms.submitForm
 import io.ktor.client.request.header
 import io.ktor.client.request.url
+import io.ktor.client.statement.HttpResponse
 import io.ktor.http.Parameters
-import org.thechance.common.data.remote.gateway.tryToExecute
+import org.thechance.common.data.remote.gateway.containsErrors
+import org.thechance.common.data.remote.gateway.getOrEmpty
 import org.thechance.common.data.remote.model.ServerResponse
 import org.thechance.common.data.remote.model.UserTokensRemoteDto
 import org.thechance.common.domain.entity.DataWrapper
+import org.thechance.common.domain.entity.InvalidCredentialsException
+import org.thechance.common.domain.entity.NoInternetException
+import org.thechance.common.domain.entity.UnknownErrorException
 import org.thechance.common.domain.entity.User
+import org.thechance.common.domain.entity.UserNotFoundException
+import java.net.ConnectException
 
 class UsersGateway(private val client: HttpClient) : IUsersGateway {
     override suspend fun getUserData(): String = "aaaa"
@@ -34,5 +43,41 @@ class UsersGateway(private val client: HttpClient) : IUsersGateway {
         }.value
 
         return Pair(result?.accessToken ?: "", result?.refreshToken ?: "")
+    }
+
+    private suspend inline fun <reified T> tryToExecute(
+        client: HttpClient,
+        method: HttpClient.() -> HttpResponse,
+    ): T {
+        try {
+            return client.method().body()
+        } catch (e: ClientRequestException) {
+            val errorMessages = e.response.body<ServerResponse<*>>().status.errorMessages
+            errorMessages?.let { throwMatchingException(it) }
+            throw UnknownErrorException()
+        } catch (e: ConnectException) {
+            throw NoInternetException()
+        } catch (e: Exception) {
+            throw UnknownErrorException()
+        }
+    }
+
+    private fun throwMatchingException(errorMessages: Map<String, String>) {
+        errorMessages.let {
+            if (it.containsErrors(UsersGateway.WRONG_PASSWORD)) {
+                throw InvalidCredentialsException(it.getOrEmpty(UsersGateway.WRONG_PASSWORD))
+            } else {
+                if (it.containsErrors(UsersGateway.USER_NOT_EXIST)) {
+                    throw UserNotFoundException(it.getOrEmpty(UsersGateway.USER_NOT_EXIST))
+                } else {
+                    throw UnknownErrorException()
+                }
+            }
+        }
+    }
+
+    companion object {
+        const val WRONG_PASSWORD = "1013"
+        const val USER_NOT_EXIST = "1043"
     }
 }
